re-written jcsync, pubsub changes that allows to run more than one instance of pubsub in one JVM, some fixes in p2pp

+++ sources/pubsub/src/main/java/pl/edu/pjwstk/mteam/pubsub/accesscontrol/AccessControlRules.java
@@ -59,7 +59,7 @@ public class AccessControlRules extends RuleSet{
 		ByteArrayInputStream istr = new ByteArrayInputStream(bytes);
 		DataInputStream dtstr = new DataInputStream(istr);
 		
-		rules = new Hashtable<Byte, Rule>();
+		rules = new Hashtable<Short, Rule>();
 		
 		try {
 			int ruleNum = dtstr.readInt();
